# llm-service/docs/contracts/llm-service.openapi.yaml
openapi: 3.1.0
info:
  title: LLM Service HTTP API
  version: 2.0.0
  license: {name: UNLICENSED, identifier: UNLICENSED}
  description: Contrato HTTP canónico de la implementación actual. Evalúa uso de IA; no expone corrección académica ni soluciones esperadas.
servers:
  - url: /api/llm
security:
  - serviceJwt: []
paths:
...
  /courses/{courseId}/golden-sets/copy-from-base/{baseVersionId}:
    post:
      summary: Copy Published Platform Golden Set into Course
      operationId: copyGoldenSetBaseToCourse
      parameters: [{$ref: '#/components/parameters/CourseId'}, {name: baseVersionId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/GoldenSetVersion'}, '403': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-sets:
    get:
      summary: List Golden Sets
      operationId: listGoldenSets
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}]
      responses: {'200': {$ref: '#/components/responses/CourseGoldenSetPage'}, '403': {$ref: '#/components/responses/Problem'}}
    post:
      summary: Create Golden Set Draft
      operationId: createGoldenSetDraft
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/GoldenSetInput'}
      responses: {'201': {$ref: '#/components/responses/GoldenSetVersion'}, '403': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-sets/{goldenSetVersionId}/cases:
    post:
      summary: Create Golden Set Case
      operationId: createGoldenSetCase
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/GoldenSetVersionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/GoldenSetCaseInput'}
      responses: {'201': {$ref: '#/components/responses/GoldenSetCaseSummary'}, '409': {$ref: '#/components/responses/Problem'}, '422': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-sets/{goldenSetVersionId}/publish:
    post:
      summary: Publish Golden Set Version
      operationId: publishGoldenSetVersion
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/GoldenSetVersionId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {$ref: '#/components/responses/GoldenSetVersion'}, '403': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
...
  /courses/{courseId}/golden-sets/{goldenSetVersionId}/cases/{caseId}:
    patch:
      summary: Update a case in a draft Golden Set
      operationId: updateGoldenSetCase
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/GoldenSetVersionId'}, {name: caseId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {$ref: '#/components/requestBodies/GoldenSetCaseInput'}
      responses: {'200': {$ref: '#/components/responses/GoldenSetCaseSummary'}, '403': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-sets/{goldenSetVersionId}:
    get:
      summary: Get Golden Set Version Detail
      operationId: getGoldenSetVersion
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/GoldenSetVersionId'}]
      responses: {'200': {$ref: '#/components/responses/GoldenSetVersion'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
    delete:
      summary: Logically delete an unused draft Golden Set family
      operationId: deleteUnusedGoldenSet
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/GoldenSetVersionId'}]
      responses: {'204': {description: Golden Set logically deleted}, '403': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
...
  /courses/{courseId}/calibrations/{runId}:
    get:
      summary: Get Calibration Run
      operationId: getCalibrationRun
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/RunId'}]
      responses: {'200': {$ref: '#/components/responses/CalibrationRun'}, '403': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/calibrations/stability-groups:
    get:
      summary: List stability groups for a course
      operationId: listCalibrationStabilityGroups
      parameters: [{$ref: '#/components/parameters/CourseId'}]
      responses: {'200': {$ref: '#/components/responses/ApiPage'}, '403': {$ref: '#/components/responses/Problem'}}
...
  /courses/{courseId}/active-calibration:
    get:
      summary: Get Active Calibration
      operationId: getActiveCalibration
      parameters: [{$ref: '#/components/parameters/CourseId'}]
      responses: {'200': {$ref: '#/components/responses/ActiveCalibration'}, '404': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/challenge-calibration-assignments:
    get:
      summary: List Challenge Calibration Assignments
      operationId: listChallengeCalibrationAssignments
      parameters: [{$ref: '#/components/parameters/CourseId'}]
      responses: {'200': {$ref: '#/components/responses/AssignmentPage'}, '403': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/pending-evaluations:
    get:
      summary: List pending evaluations
      operationId: listPendingEvaluations
      parameters: [{$ref: '#/components/parameters/CourseId'}]
      responses: {'200': {$ref: '#/components/responses/PendingEvaluationPage'}, '403': {$ref: '#/components/responses/Problem'}}
...
  /admin/institutional-calibration/runs/{runId}:
    get:
      summary: Get an institutional calibration run with its per-dimension errors
      operationId: getInstitutionalCalibrationRun
      parameters: [{name: runId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {$ref: '#/components/responses/ApiObject'}, '403': {$ref: '#/components/responses/Problem'}, '404': {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    serviceJwt: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
...
    CourseId: {name: courseId, in: path, required: true, schema: {type: string, format: uuid}}
...
    GoldenSetVersionId: {name: goldenSetVersionId, in: path, required: true, schema: {type: string, format: uuid}}
...
    RunId: {name: runId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
    Size: {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  requestBodies:
...
    GoldenSetInput: {required: true, content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string}, copyFromBaseVersionId: {type: string, format: uuid}}}}}}
    GoldenSetCaseInput: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GoldenSetCaseInput'}}}}
...
  responses:
    Problem: {description: RFC 7807 Problem Details}
...
    GoldenSetVersion: {description: Golden Set, content: {application/json: {schema: {$ref: '#/components/schemas/GoldenSetVersion'}}}}
...
    GoldenSetCaseSummary: {description: Resumen del caso creado, content: {application/json: {schema: {$ref: '#/components/schemas/GoldenSetCaseSummary'}}}}
...
    CalibrationRun: {description: Corrida, content: {application/json: {schema: {$ref: '#/components/schemas/CalibrationRun'}}}}
...
    ActiveCalibration: {description: Activa, content: {application/json: {schema: {$ref: '#/components/schemas/ActiveCalibration'}}}}
...
    CourseGoldenSetPage: {description: Resumen de versiones y casos del curso, content: {application/json: {schema: {$ref: '#/components/schemas/CourseGoldenSetPage'}}}}
...
    AssignmentPage: {description: Página, content: {application/json: {schema: {$ref: '#/components/schemas/AssignmentPage'}}}}
...
    PendingEvaluationPage: {description: Página, content: {application/json: {schema: {$ref: '#/components/schemas/PendingEvaluationPage'}}}}
...
    ApiPage: {description: Página de recursos cuyo detalle se mantiene en el controller, content: {application/json: {schema: {$ref: '#/components/schemas/ApiPage'}}}}
    ApiObject: {description: Recurso administrativo; sus campos concretos se describen en la operación y en el consumidor de referencia, content: {application/json: {schema: {$ref: '#/components/schemas/ApiObject'}}}}
  schemas:
    Dimensions:
      type: object
      required: [AUTONOMY, CLARITY, PROGRESSION, COMPLIANCE, EFFICIENCY]
      properties:
        AUTONOMY: {type: integer, minimum: 0, maximum: 100}
        CLARITY: {type: integer, minimum: 0, maximum: 100}
        PROGRESSION: {type: integer, minimum: 0, maximum: 100}
        COMPLIANCE: {type: integer, minimum: 0, maximum: 100}
        EFFICIENCY: {type: integer, minimum: 0, maximum: 100}
...
    TranscriptMessage:
      type: object
      required: [role, content, position]
      properties: {role: {type: string, enum: [STUDENT, TUTOR]}, content: {type: string, minLength: 1}, position: {type: integer, minimum: 0}, occurredAt: {type: string, format: date-time}}
    GoldenSetCaseInput:
      type: object
      required: [transcript, referenceScores]
      properties: {transcript: {type: array, minItems: 1, items: {$ref: '#/components/schemas/TranscriptMessage'}}, challengeContext: {type: object, required: [externalChallengeId, statement], properties: {externalChallengeId: {type: string}, statement: {type: string}}}, metadata: {type: object, additionalProperties: true}, author: {type: string, deprecated: true, description: 'Ignorado: el servidor registra al docente autenticado.'}, referenceScores: {$ref: '#/components/schemas/Dimensions'}, scoreJustifications: {type: object, additionalProperties: {type: string}}}
    GoldenSetCase:
      allOf: [{$ref: '#/components/schemas/GoldenSetCaseInput'}, {type: object, required: [id, goldenSetVersionId, order, reviewState], properties: {id: {type: string, format: uuid}, goldenSetVersionId: {type: string, format: uuid}, order: {type: integer}, reviewState: {type: string, enum: [DRAFT, REVIEWED]}}}]
    GoldenSetVersion:
      type: object
      required: [id, familyId, version, state, name]
      properties: {id: {type: string, format: uuid}, familyId: {type: string, format: uuid}, version: {type: integer}, state: {type: string, enum: [DRAFT, PUBLISHED, SUPERSEDED]}, name: {type: string}, cases: {type: array, items: {$ref: '#/components/schemas/GoldenSetCase'}}}
    GoldenSetCaseSummary:
      type: object
      required: [id, order, author, reviewState]
      properties: {id: {type: string, format: uuid}, order: {type: integer}, author: {type: string}, reviewState: {type: string, enum: [DRAFT, REVIEWED]}}
    CourseGoldenSetVersion:
      type: object
      required: [id, familyId, name, version, state, cases]
      properties: {id: {type: string, format: uuid}, familyId: {type: string, format: uuid}, name: {type: string}, version: {type: integer}, state: {type: string, enum: [DRAFT, PUBLISHED, SUPERSEDED]}, basedOnVersionId: {type: string, format: uuid}, cases: {type: array, items: {$ref: '#/components/schemas/GoldenSetCaseSummary'}}}
...
    CalibrationCaseResult:
      type: object
      required: [caseId, modelScores, humanFinalScore, modelFinalScore, dimensionErrors, finalError]
      properties: {caseId: {type: string, format: uuid}, modelScores: {$ref: '#/components/schemas/Dimensions'}, humanFinalScore: {type: number}, modelFinalScore: {type: number}, dimensionErrors: {$ref: '#/components/schemas/Dimensions'}, finalError: {type: number}}
    CalibrationRun:
      type: object
      required: [id, state, rubricVersionId, goldenSetVersionId, modelDeploymentId, progress]
      properties: {id: {type: string, format: uuid}, state: {type: string, enum: [QUEUED, RUNNING, PASSED, FAILED, CANCELLED]}, rubricVersionId: {type: string, format: uuid}, goldenSetVersionId: {type: string, format: uuid}, modelDeploymentId: {type: string, format: uuid}, progress: {type: integer, minimum: 0, maximum: 100}, maeFinal: {type: number}, maxIndividualError: {type: integer}, results: {type: array, items: {$ref: '#/components/schemas/CalibrationCaseResult'}}}
...
    ActiveCalibration: {type: object, required: [courseId, calibrationRunId, activatedAt], properties: {courseId: {type: string, format: uuid}, calibrationRunId: {type: string, format: uuid}, activatedAt: {type: string, format: date-time}}}
...
    ChallengeCalibrationAssignment: {type: object, required: [challengeId, calibrationRunId, locked], properties: {challengeId: {type: string, format: uuid}, calibrationRunId: {type: string, format: uuid}, locked: {type: boolean}, lockedAt: {type: string, format: date-time}}}
...
    PendingEvaluation:
      type: object
      required: [id, attemptId, assignmentId, calibrationRunId, state, reason, queuedAt]
      properties: {id: {type: string, format: uuid}, attemptId: {type: string, format: uuid}, assignmentId: {type: string, format: uuid}, calibrationRunId: {type: string, format: uuid, description: 'Immutable calibration version applied to the attempt'}, state: {type: string, enum: [QUEUED, RUNNING, COMPLETED, FAILED]}, reason: {type: string, enum: [CALIBRATION_UNAVAILABLE]}, queuedAt: {type: string, format: date-time}}
...
    CourseGoldenSetPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/CourseGoldenSetVersion'}}}}
...
    AssignmentPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/ChallengeCalibrationAssignment'}}}}
...
    PendingEvaluationPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/PendingEvaluation'}}}}
    ApiPage: {type: object, required: [items], properties: {items: {type: array, items: {type: object, additionalProperties: true}}}}
    ApiObject: {type: object, additionalProperties: true}
...